Move correlated GP to exact submodule
